// src/models/pointUpdates.ts
export interface IPointUpdates {
  // Ссылка на получение последних обновлений баллов для контрольной точки
  url: string;
  // Дата обновления баллов контрольной точки в формате ДД.ММ.ГГГГ
  date: string;
}